wip: refactor structure
