+++ internal/interrupt/context.go
@@ -1,0 +1,46 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package interrupt
+
+import (
+	"context"
+	"os"
+	"os/signal"
+	"syscall"
+)
+
+// Context returns a context that is canceled on SIGINT and SIGTERM.
+func Context() (context.Context, func()) {
+	return WrappedContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
+}
+
+// WrappedContext returns a new context wrapping the provided context and
+// canceling it on the provided signals.
+func WrappedContext(ctx context.Context, signals ...os.Signal) (context.Context, func()) {
+	ctx, closer := context.WithCancel(ctx)
+
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, signals...)
+
+	go func() {
+		select {
+		case <-c:
+			closer()
+		case <-ctx.Done():
+		}
+	}()
+
+	return ctx, closer
+}
