+++ internal/kubernetes/kv_maps/abstract_kv_map.go
@@ -1,4 +1,4 @@
-package kubernetes
+package kv_maps
 
 import (
 	"reflect"
