+++ internal/kubernetes/pod_controllers/statefulsets.go
@@ -1,9 +1,36 @@
-package kubernetes
+package pod_controllers
 
 import (
+	"fmt"
+
 	v1 "k8s.io/api/apps/v1"
-	v12 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+
+	"k8s-cluster-comparator/internal/config"
+	"k8s-cluster-comparator/internal/logging"
 )
+
+func CompareStateFulSets(clientSet1, clientSet2 kubernetes.Interface, namespace string, skipEntities config.SkipEntitiesList) (bool, error) {
+	var (
+		isClustersDiffer bool
+	)
+	statefulSet1, err := clientSet1.AppsV1().StatefulSets(namespace).List(metav1.ListOptions{})
+	if err != nil {
+		return false, fmt.Errorf("cannot obtain statefulsets list from 1st cluster: %w", err)
+	}
+
+	statefulSet2, err := clientSet2.AppsV1().StatefulSets(namespace).List(metav1.ListOptions{})
+	if err != nil {
+		return false, fmt.Errorf("cannot obtain statefulsets list from 2nd cluster: %w", err)
+	}
+
+	apc1List, map1, apc2List, map2 := PrepareStatefulSetMaps(statefulSet1, statefulSet2)
+
+	isClustersDiffer = ComparePodControllerSpecs(map1, map2, apc1List, apc2List, namespace)
+
+	return isClustersDiffer, nil
+}
 
 // PrepareStatefulSetMaps prepares StatefulSet maps for comparison
 func PrepareStatefulSetMaps(obj1, obj2 *v1.StatefulSetList) ([]AbstractPodController, map[string]IsAlreadyComparedFlag, []AbstractPodController, map[string]IsAlreadyComparedFlag) { //nolint:gocritic,unused
@@ -19,6 +46,7 @@ func PrepareStatefulSetMaps(obj1, obj2 *v1.StatefulSetList) ([]AbstractPodContro
 
 	for index, value := range obj1.Items {
 		if _, ok := ToSkipEntities[ObjectKindWrapper(value.Kind)][value.Name]; ok {
+			logging.Log.Debugf("statefulset %s is skipped from comparison due to its name", value.Name)
 			continue
 		}
 
@@ -27,7 +55,7 @@ func PrepareStatefulSetMaps(obj1, obj2 *v1.StatefulSetList) ([]AbstractPodContro
 
 		apc1List = append(apc1List, AbstractPodController{
 			Metadata: AbstractObjectMetadata{
-				Type: v12.TypeMeta{
+				Type: metav1.TypeMeta{
 					Kind:       "statefulsets",
 					APIVersion: "apps/v1",
 				},
@@ -44,6 +72,7 @@ func PrepareStatefulSetMaps(obj1, obj2 *v1.StatefulSetList) ([]AbstractPodContro
 
 	for index, value := range obj2.Items {
 		if _, ok := ToSkipEntities[ObjectKindWrapper(value.Kind)][value.Name]; ok {
+			logging.Log.Debugf("statefulset %s is skipped from comparison due to its name", value.Name)
 			continue
 		}
 		indexCheck.Index = index
@@ -51,7 +80,7 @@ func PrepareStatefulSetMaps(obj1, obj2 *v1.StatefulSetList) ([]AbstractPodContro
 
 		apc2List = append(apc2List, AbstractPodController{
 			Metadata: AbstractObjectMetadata{
-				Type: v12.TypeMeta{
+				Type: metav1.TypeMeta{
 					Kind:       "statefulsets",
 					APIVersion: "apps/v1",
 				},
